market master box weight data type changes

+++ src/main/java/com/sericulture/registration/model/entity/MarketMaster.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Entity(name = "market_master")
 @AllArgsConstructor
@@ -30,7 +31,7 @@ public class MarketMaster extends BaseEntity implements Serializable {
     private String marketMasterAddress;
 
     @Column(name = "box_weight")
-    private Long boxWeight;
+    private BigDecimal boxWeight;
 
     @Column(name = "lot_weight")
     private Long lotWeight;
